Kill the XHPJS class
It was a holder for functions and didn't need to be a class.
This code was written in a time when autoloading functions
was something that the default autoloader (composer) couldn't do.
I moved these pseudo contructors the member functions.
to_js_value (XHPJS::MapArgument) doesn't look like a public API,
so I movesed that to _Private.